flytta funksjoner til andre filer
