+++ referanse-data/elhub/party_gln_eic_mga.csv
@@ -1,4 +1,4 @@
-"PARTY_NAME, GLN, GRID_AREA_ID"
+"PARTY_NAME","GLN","GRID_AREA_ID","GRID_AREA_NAME"
 "Vestmar Nett AS","7080005054928","50Y-7FS806KF24QR","KRAGERØ1"
 "VENI Metering AS","7080005050616","50Y0B-1DGY7ZN8KB","SN01HAFSL1"
 "VENI Metering AS","7080005050616","50Y2CKL12AOFY8UC","SN01BKKN1"
